+++ ikarus/python/finiteelements/registerferequirements.hh
@@ -5,7 +5,7 @@
 
 #include <dune/python/pybind11/pybind11.h>
 
-#include <ikarus/python/finiteelements/valuewrapper.hh>
+#include <ikarus/python/finiteelements/scalarwrapper.hh>
 
 namespace Ikarus::Python {
 
@@ -37,7 +37,8 @@ void registerFERequirement(pybind11::handle scope, pybind11::class_<FE, options.
         "globalSolution", [](FERequirements& self) { return self.globalSolution(); },
         pybind11::return_value_policy::reference_internal);
     lv.def(
-        "insertParameter", [](FERequirements& self, ValueWrapper<double>& parVal) { self.insertParameter(parVal.val); },
+        "insertParameter",
+        [](FERequirements& self, ScalarWrapper<double>& parVal) { self.insertParameter(parVal.value()); },
         pybind11::keep_alive<1, 2>(), "parameterValue"_a.noconvert());
 
     lv.def("parameter", [](const FERequirements& self) { return self.parameter(); });